pyproject.toml
[tool.poetry]
name = "oca-nats-config"
version = "0.1.0"
description = ""
authors = ["elasek <elasek@akond.com>"]
readme = "README.md"
packages = [{include = "oca_nats_config"}]

[tool.poetry.dependencies]
python = "^3.10"
nats-py = "^2.3.1"
serverish = {path = "../serverish", develop = true}
confuse = ">=1.7.0,<1.8.0"


[tool.poetry.scripts]
build = "oca_nats_config.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
